feat(cognitive-loop): Round 7 — exploration/threshold ProposalCollector migration, phase tests, telemetry
Complete feedback dual-write migration: exploration_urge (27 sites) and
adaptive_threshold_scale (6 sites) now use ProposalCollector attribution
with descriptive source labels. Add dynamics phase coverage tests,
consensus-vs-sequential A/B test, bath phase accessors, and fix f64 type
in telemetry profile comparison.

+++ tests/telemetry_validation.rs
@@ -352,11 +352,11 @@ fn profile_comparison_full_vs_default() {
     let full_max_cl = full_results
         .iter()
         .map(|r| r.metadata.consciousness_level)
-        .fold(0.0f32, f32::max);
+        .fold(0.0f64, f64::max);
     let default_max_cl = default_results
         .iter()
         .map(|r| r.metadata.consciousness_level)
-        .fold(0.0f32, f32::max);
+        .fold(0.0f64, f64::max);
 
     assert!(
         full_max_cl >= default_max_cl,
